Terrain Threading and Small Changes
Fixed the screenshot script to be able to take multiple screenshots and
modified gitIgnore to ignore the folder where they are stored.
The terrain is not threaded as you travel from the staring area.
This allows for much larger play area's to be used without hassel.
Terrain also has a global scale map. currently this is hard coded but
will add more options in a future commit.